refactor: 5xx 에러 알림 관련 인터페이스 분리
