+++ src/main/java/com/koddy/server/global/exception/notify/ErrorNotifier.kt
@@ -1,0 +1,11 @@
+package com.koddy.server.global.exception.notify
+
+import jakarta.servlet.http.HttpServletRequest
+
+@FunctionalInterface
+fun interface ErrorNotifier {
+    fun send(
+        request: HttpServletRequest,
+        exception: Exception,
+    )
+}
